fixed naming convention. Moved useWindowDimensions from /components/ -> /hooks/
