+++ src/App.jsx
@@ -8,9 +8,9 @@ function App() {
       <BrowserRouter>
         <Routes>
           <Route path='/' element={<Layout><LandingPage /></Layout>} />
-          <Route path='/Research' element={<Layout><Research /></Layout>} />
-          <Route path='/Publications' element={<Layout><Publications /></Layout>} />
-          <Route path='/Team' element={<Layout><Team /></Layout>} />
+          <Route path='/research' element={<Layout><Research /></Layout>} />
+          <Route path='/publications' element={<Layout><Publications /></Layout>} />
+          <Route path='/team' element={<Layout><Team /></Layout>} />
           <Route path='*' element={<NotFound />}></Route>
         </Routes>
       </BrowserRouter >
