+++ src/components/index.js
@@ -1,6 +1,6 @@
-export { default as LandingPage } from './LandingPage.jsx';
-export { default as Research } from './Research.jsx';
-export { default as Publications } from './Publications.jsx';
-export { default as Team } from './Team.jsx';
-export { default as NotFound } from './NotFound.jsx';
-export { default as Layout } from './Layout.jsx';
+export { default as LandingPage } from './landingPage.jsx';
+export { default as Research } from './research.jsx';
+export { default as Publications } from './publications.jsx';
+export { default as Team } from './team.jsx';
+export { default as NotFound } from './notFound.jsx';
+export { default as Layout } from './layout.jsx';
