+++ src/styles/Publications.css
@@ -69,7 +69,7 @@
 
 @media (max-width: 1079px) {
     .publication-title{
-        padding-top: 20px;
+        padding-top: 30px;
     }
 
 }
