+++ src/styles/Research.css
@@ -59,7 +59,7 @@ h4 {
     .research {
         flex-direction: column;
         align-items: center;
-        margin-top: 20px;
+        margin-top: 30px;
     }
 
     .research-panel {
