+++ src/styles/Team.css
@@ -65,6 +65,6 @@
 
 @media (max-width: 1079px) {
     .team-title {
-        padding-top: 20px;
+        padding-top: 30px;
     }
 }
